// interfaces/location.interface.ts
export interface Location {
    type:        string;
    query:       string[];
    features:    Feature[];
    attribution: string;
}

export interface Feature {
    id:         string;
    type:       string;
    place_type: string[];
    relevance:  number;
    properties: Properties;
    text:       string;
    place_name: string;
    bbox:       number[];
    center:     number[];
    geometry:   Geometry;
    context:    Context[];
}

export interface Context {
    id:         string;
    short_code: string;
    wikidata:   string;
    text:       string;
}

export interface Geometry {
    type:        string;
    coordinates: number[];
}

export interface Properties {
    wikidata:    string;
    short_code?: string;
}

export interface InputsProps{
    pickup: string;
    dropoff: string;
}